plugins/source/square/requirements.txt
black==23.10.1
cloudquery-plugin-sdk==0.1.8
pyarrow==14.0.1
squareup==29.0.0.20230720
pytest==7.4.3
